requirements.txt
openai==1.23.3
argparse==1.4.0
sphinx
myst_parser
jsonlines
sphinx_rtd_theme
pyyaml
google-generativeai
huggingface_hub
enlighten